kursach3sem/requirements.txt
asgiref==3.8.1
astroid==3.3.8
diff-match-patch==20241021
dill==0.3.9
Django==5.1.4
django-filter==24.3
django-import-export==4.3.3
django-simple-history==3.7.0
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
pylint==3.3.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
sqlparse==0.5.3
tablib==3.7.0
tomlkit==0.13.2
openpyxl==3.1.5
django-widget-tweaks==1.5.0
